Non gnome stuff depends on this lib so people who don't have gnome installed don't have the gnome-env package installed. Ergo thier /etc/ld.so.conf doesn't have  /opt/gnome/lib in it. So a package like enlightenment which depends on libghttp correctly installs the lib, can't find it when run because the linker cache is never updated.
Probably this library need to be moved out of gnome-base and be
reconfigured to install to /usr/lib since it's not _JUST_ gnome specific.
When I do that, I'll also need to update all the DEPENDS and RDEPEPENDS in
the tree as well. I'll do this this weekend in my copious spare time :-)