Added two packages the first is hogwash which is a layer2 packet scrubber. It lives right on top of the network driver and will drop or sanitize packets based on signature. The authors installed it on an unpatched RH-6.2 box and entered it into the capture the flag competition at defcon 9 and the box came back unscathed. I've included the rules file too. as well as some documentation I copied off thier website. I may write a manpage for them too.
The second is a supporting library for it that provides an API for commonly
used low-level network functions (mainly packet injection).
Used by packet scrubbers and the like, not to be confused with the perl
libnet.

They hogwash folks have a how-to on thier site for setting up an packet scrubber with no
IP stack to protect your network, sortof a stealth firewall. Since it never
modifies the mac address it's nearly impossible to detect and attack. Will
prevent spoofing of IPs as well as 95% of the common attacks.