Wow. Didnt notice this *very* cool logger before.. I think I just found my next logging daemon. Has regexp support and stuff. Im glad I got to take a look at this.. neato.  :>~
Note that I removed the "-s" option from the initscript, this
stuff is better left up to the admin of the box. Synchronous
mode could be detrimental to performance on a loaded machine..
I think we'd agree that the distribution need not dictate
policy. We *should* provide a solid init *system*.. which
is all the competent admin needs. ;) I honestly feel that
users are just better off pointed to the docs, rather have
having a configured system decided upon for them.