Version bump.
"xCHM is now the first non-MS CHM viewer able to parse and use the binary table
of contents and index files inside CHM books. Upgrading to 1.11 is strongly
recommended, as this version loads CHMs way faster, and it also fixes a couple
of bugs."

Cleaned up the ebuild and added an nls USE flag.  Removed old version.
(Portage version: 2.1.2.2)
 (Unsigned Manifest commit)